+++ tests/Unit/Calculators/Dominion/MilitaryCalculatorTest.php
@@ -572,11 +572,10 @@ class MilitaryCalculatorTest extends AbstractBrowserKitTestCase
 
         $this->landCalculator->shouldReceive('getTotalLand')->with($dominion)->andReturn(1000)->byDefault();
         $dominion->shouldReceive('getTechPerkValue')->with('wizard_strength_recovery')->andReturn(1)->byDefault();
-        $dominion->shouldReceive('getCourtSeat')->andReturn('mage')->byDefault();
         $dominion->shouldReceive('isActive')->andReturn(true)->byDefault();
         $dominion->shouldReceive('getAttribute')->with('building_wizard_guild')->andReturn(10)->byDefault();
         $dominion->shouldReceive('getAttribute')->with('wizard_strength')->andReturn(50)->byDefault();
 
-        $this->assertEquals(5.3, $this->sut->getWizardStrengthRegen($dominion));
+        $this->assertEquals(5.1, $this->sut->getWizardStrengthRegen($dominion));
     }
 }