+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes relevant to players in this project will be documented in th
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 
 ## [Unreleased]
+
+## [1.39.0] - 2024-03-15
 ### Added
 - New Government Role: Spymaster
   - Can cancel any bounty
@@ -2118,7 +2120,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 ### Added
 - This CHANGELOG file
 
-[Unreleased]: https://github.com/OpenDominion/OpenDominion/compare/1.38.4...HEAD
+[Unreleased]: https://github.com/OpenDominion/OpenDominion/compare/1.39.0...HEAD
+[1.39.0]: https://github.com/OpenDominion/OpenDominion/compare/1.38.4...1.39.0
 [1.38.4]: https://github.com/OpenDominion/OpenDominion/compare/1.38.3...1.38.4
 [1.38.3]: https://github.com/OpenDominion/OpenDominion/compare/1.38.2...1.38.3
 [1.38.2]: https://github.com/OpenDominion/OpenDominion/compare/1.38.1...1.38.2